requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
better-profanity==0.7.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
dj-database-url==2.3.0
Django==5.0.7
django-compat==1.0.15
djangorestframework==3.15.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.65.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
hyperlink==21.0.0
idna==3.7
imageio==2.34.2
incremental==22.10.0
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opennsfw2==0.13.7
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scipy==1.14.0
service-identity==24.1.0
setuptools==71.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tifffile==2024.7.2
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.34.0
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.8.2
wrapt==1.16.0
zope.interface==6.4.post2
